move LootModifierManager instance, getter & reloader
